Extract common OAuth 2 functionality in separate class
